Added the Required and RequiredErrorMessage attributes to each of the necessary items on the Add Person/Visitor modal.
Added a CSS class for the datepicker to add the same amount of padding underneath it as the text boxes and drop down lists have. This was so the required error messages all lined up across the page. I also added another class to style the error message text to be red. Like an error should be.

Fixed an issue with adding a new person where the if you chose an ability level on the new person it would cause an error. Had to get a full person object and load the attributes first. Hat tip to the AbilityLevelSelect page for this bit of code!