Changed the Admin page to clear the "active" class on all the buttons in the repeater when none of them are selected and you click the "Ok" button. Fixes an issue where when none of them were selected, and the modal fired, it would put the active class back on any button that was selected originally, but it wouldn't set the hidden field value, so what you saw on the screen didn't match what was actually selected.
Fixed Add Person/Visitor panel to clear out the SelectedDate property of the date picker when opening the panel. If there was a date
previously entered in the date picker, it returned even though we were clearing out the text.

Styling change on the Add Family modal to give the rows a little bottom margin to spread things out. Also changed the Save button on the
Add Family modal to set CausesValidation = false. Without this the button click wouldn't process.

Added a LoadGroupTypes function to Family Select that will load the group types to the CheckInPerson for anyone being added as a person or
visitor through the modal. The group types for people added this way were not being loaded, so they wouldn't get a room calculated for
them. Therefore they wouldn't show up on the confirmation page.

Changed FilterGroupsByAbilityLevel to do ToUpper() on the personAbilityLevel guid.

Changed SelectByBestFit to check for the gradeGroups and ageGroups having a count > 0 before trying to set something. Was getting an error when it tried to calculate the appropriate groupMatchGrade when the gradeGroup didn't have anything in it.

Fixed a bug in Family Select where if you came back from the confirmation page it showed any selected visitors as selected but in reality they weren't. Added code to add these selected visitors id's to the hfSelectedVisitor hidden field so that they would be correctly processed when the Next button is clicked.

Fixed the Activity Select page so that when the last item in the Selected grid is deleted, and someone clicks the Back button, those changes aren't saved and everything is set back correctly.