Upload / Timeout Bug Fixed
A new method for IFieldType was required in order to allow field types
to retrieve values on a pre-existing context which was already being
used to save a new entity.  The use of the existing context with
associated transaction eliminated a deadlock error.  The
BinaryFileFieldType class implements the new option.  Others could be as
well if the error is seen in other circumstances.