Fix bug when updating a BinaryFile to use a different storage provider
When a BinaryFile is saved using a BinaryFileType with Database storage, and then the file is updated to a BinaryFileType with FileSystem storage, the BinaryFile table gets updated in some places, but the file does not actually get saved where it should.  Saving the contents of the BinaryFile before deleting and then setting the contents back again correctly populates the data, and sets the dirty flag so it gets saved properly.