+ Set best practice for default property
Reference [here](https://weblogs.asp.net/scottgu/442448) and [here](https://bitsandscrews.com/10-solid-tips-to-increase-and-optimize-iis-performance-for-2018-covers-asp-net-wordpresscoldfusion-and-sharepoint/#Disable_ASPNet_debugging_from_WebConfig).  This setting is the default, which gets passed to the installer and deployed to production environments.

To fix existing Rock installations, we may need to add a setting under System Settings > System Config.