Create Person_BirthDate.sql
The current BirthDate column in the Person table is not able to be persisted.  This new calculation method not only mirrors the existing computations (and uses slightly less CPU) but allows for it to be persisted which enables the ability for indexes as well as resulting in a speed increase from 2000ms to 500ms at scale.