Final Commit
Migration files completed.

Things still needed to change:
Workflow migration didnt make the Connection type for the Volunteer Interest Form Connection Opportunity attribute. Supposed to be default of Kid Ministry.

The Connection Type filter on /GenerateNCReport was unable to be migrated over. Navigate to the page, select the gear icon on the block, and add a filter of Name: Connection Type and Type: Connection Type.

The file path for bootstrap inside the lava to pdf block for the adult and student application workflow will need to be changed.

The file path for the image used in the name clearing report will need to be checked. This is located on /NCReport, the Internal Application Root Global Attribute might need to be properly setup for it to work.