Make showing a schedules name instead of time configurable
Added the ability to either display a Schedules Time or Name with a boolean block attribute.

In the rest of the scheduling process you deal with schedules by name this is also how we deal with them internally. A person rarely sees the actual time a schedule starts instead they see Sunday 1st, Sunday 2nd, Sunday 3rd etc... All of these schedules start at 5:30am, for us and for multiple reasons, which might be unique to us. This is the only place I have found where it's an actual time displayed and not the schedule name, and for us it's causing issues. This is Not a Bug, but when a volunteer looks at their schedule knowing that they've accepted for 3rd service and then they see 5:30 am they freak out.