:sparkles: Group Schedule Toolbox
Added the ability to have Schedule names displayed to the user via a block setting.

In most of the group scheduling process you deal with schedules by name this is also how we deal with them internally. A person rarely sees the actual time a schedule starts instead they see Sunday 1st, Sunday 2nd, Sunday 3rd etc... All of these schedules start at 5:30am, for us and for multiple reasons, which might be unique to us. This is the only place I have found where it's an actual time displayed and not the schedule name, and for us it's causing issues. 

This is Not a Bug, but when a volunteer looks at their schedule, after they've confirmed, knowing that they've accepted for Sunday 3rd service and then they only see 5:30am they freak out. No where does it show the schedule name. Now you can change the behavior in the block settings.