Fixes Inactive Opportunity Types From Displaying
This is related to issue #3946 

If you have a Connection Opportunity Search page and disable that Opportunity type The page will still display all of the opportunities and let you sign up. I've submitted a pr for the signup page already. Added a block option for toggling warning text. In our Church's use case I wouldn't want the warning text as I have several of these blocks on the same page and simply want it to not show anything. Default is show the warning.
 
I didn't make a separate issue to track this