Project Init
This is the code of the project at its current form in development on 10/29/2020.  I parsed the generated index as a lava file with comments including the head link tabs.  I think there would be a way to have webpack generate this file, but we would need a more robust add link to head tag to handle the prefetch as.