Login Block with Allow Guest
This is code from their old custom block.  It adds an option to enable Continue as Guest when the return url contains registraiton or registration instance.  The registraiton page as a redirect on the page that checks for login or Guest=True.  This was set up as a way to suggest login/account creation, but not require it.