Update Public String PersonActionIdentifier
I'm not sure if this is exactly how you would want to fix this, but the page is looking for "PersonActionIdentifier" in the URL as a page parameter but when you click on "add pledge" on the person profile it's sending the value through as "rckid". Just submitting this fix so that it works properly again.