Modified Advantage Person Profile Page
This page is a modified version or the person profile page. It's designed to support the needs of the Church Relations/Advantage pages in Rock. 

Context from Triumph.....

Why this is necessary:

This layout is an exact duplicate of the PersonProfileHome.aspx file (github link) that’s already in that directory – the only reason we need it is because there are blocks installed on the core Person Profile Home layout that we don’t want to have also show up on the advantage person profile page.

In core this layout is only used on a single page and so the blocks could, in theory, be moved to only be added to the single page. That would have allowed us to use the core layout for the Advantage profile page. But on evaluation, that seems like a bigger edit to core’s design and since your team has added another page that uses that layout (and therefore gets those blocks), we didn’t want to add the technical debt of your having to keep those pages in sync with their blocks.

So instead we copied the core layout to a different name, and we’re using that layout for your Advantage project so its blocks are independent of the regular person profile layout. If the core PersonProfileHome page is ever updated in core, this cloned layout should be updated to match.