- Fixed an issue where serve team members couldn't be edited
This fixes an issue where members of scheduled serve teams couldn't be edited on the Group Roster block on the public site.

This was caused because scheduled serve teams only allow a member to be in the group once, whereas in other groups the person can be in the group multiple times with different roles. The logic I edited was meant to make sure that the person was always at least in the Member role for groups that aren't scheduled. It did this by adding the member role to the list of roles to give the group member. This means that if the user selected the Member role, the Member role would be added to the list a second time by this logic. This would cause the validation logic to see two roles (even though they are the same role). Since scheduled serve team members can't have more than one role in the same group, the validation logic would throw an error.

This fixes that issue by only adding the Member role to the list if the group isn't a scheduled serve team.