- Fixed an Issue Where Filtered Groups Were Shown on the Map
Fixed an issue with the group locator where the map was not hidden when there were no groups to show on the map. This caused the map to be shown in whatever the previous state was. This led to groups that were filtered out being shown on the map when there were no groups found that met the filters.

This fixes that issue by adding a step after the filtered data is retrieved to hide the map if there are no results.